Turn off security for bin
This has to be done AFTER the config file has been loaded to make sure
the setting isn't overwritten.

NCBO-1297 #resolve #comment Security is now disabled when running the
cron project as a binary (it has to be done here because cron gets
loaded in ontologies_api and we can't disable security there).